{{cookiecutter.project_name}}/Backend/requirements.txt
Django==2.0.3
psycopg2==2.8.4 --no-binary psycopg2
pillow==5.3.0
django-autofixture==0.12.1
coreapi==2.3.3
django-braces==1.13.0
djangorestframework==3.8.2
django-filter==1.1.0
rest-framework-cache==0.1
djangorestframework-filters==0.10.2
django-oauth-toolkit==1.0.0
easy-thumbnails==2.5
markdown==2.6.11
simplejson==3.16.0
drf_autodocs==0.4.4
sendgrid==5.6.0
pyyaml==3.13
ua-parser==0.8.0
user-agents==1.1.0
django-user-agents==0.3.2
django-extensions==2.1.2
git+https://github.com/cdrx/django-admin-menu.git
git+https://github.com/groveco/django-sql-explorer.git#egg=django-sql-explorer
django-cors-middleware==1.3.1
git+https://github.com/django-import-export/django-import-export.git#egg=django-import-export
django-jet
django-imagekit
django-map-widgets==0.2.0
dj-database-url
python-decouple
django-image-cropping
django-tinymce4-lite
googlemaps
django-cacheops==4.1
git+https://test.pypi.org/simple/ionicdjangocrudgenerator==1.0.0
kombu==4.5.0
redis==3.2.0
django-redis==4.8.0
hiredis==0.2.0
celery==4.2.1
